session-3-ai-agents/agents/lunch-selection-agent/requirements.txt
# Lunch Selection Agent Dependencies
google-genai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
